requirements.txt
llama-cpp-python
chromadb
sentence-transformers
faster-whisper
pyttsx3
sounddevice
numpy
pypdf